Finish detangling the What4 backend from the evaluator
